@@ -4,7 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <AssemblyName>Mikodev.Binary.Benchmarks.Abstractions</AssemblyName>
     <RootNamespace>Mikodev.Binary.Benchmarks.Abstractions</RootNamespace>
-    <LangVersion>10.0</LangVersion>
+    <LangVersion>12</LangVersion>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
