@@ -41,7 +41,7 @@ public class CollectionBenchmarks
             : new VariableNativeConverter<int>() as Converter<int>;
         var generator = Generator.CreateDefaultBuilder().AddConverter(converter).Build();
 
-        this.hashSet = new HashSet<int> { 1313 };
+        this.hashSet = [1313];
         this.linkedList = new LinkedList<int>(new[] { 1313 });
         this.dictionary = new Dictionary<int, int> { [1313] = 1313 };
 
